Update logging and search methods, refine timeouts
Improved the codebase by adding more methods in the logging and search functions. The updates distribute the scraping process into smaller functions for better code management across multiple clients such as Blibli, Bukalapak, Shopee, and Tokopedia.

A critical update includes setting a timeout limit of 5000 ms (5 seconds) for requests to prevent possible blocking of threads due to network delay. This keeps execution efficient and limits the delay due to network requests.

The updates also include additional debug logs and optimized test cases to ensure better debugging and testing visibility.

In this update, changes in the ecommerce-api-endpoint project involve refining API route paths for more specific searches. These changes will ultimately lead to superior performance, better risk management, and simplify the maintenance of code.

+++ ecommerce-api-core/src/main/kotlin/com/ramusthastudio/ecommerce/model/BlibliSearchResponse.kt
@@ -783,7 +783,7 @@ data class BlibliSearchResponse(
             @SerialName("storeClosingInfo")
             val storeClosingInfo: StoreClosingInfo?,
             @SerialName("promoEndTime")
-            val promoEndTime: Int?,
+            val promoEndTime: Long?,
             @SerialName("debugData")
             val debugData: DebugData?,
             @SerialName("allCategories")
